feat(shareable-lists): paginated items on a list
Provides a cursor pagination API, but uses offsets
under the hood (makes more sense given the data model,
and should be sufficient for the sizes of lists we
intend to support [<= 1k])

Previously every item was retrieved in a single query,
which is ok in many cases (vast majority of lists have
0-100 items), but list sizes may change after the bulk
add mutation is implemented. Currently users must add
items to list manually, one at a time.

[POCKET-9550]