fix: variable input key name
Aligns variable values and input keys to reduce
referential errors. The variables were expecting
to assign a key "filters" to the input variable
"filter", but the "filters" key was passed as "filter"
(aligned with the actual graphql input object).

I'd expect that the incredibly unwieldy types would
have caught this error, but...