populate timestamp with default if not provided
Adds the current time in seconds from epoch
to the timestamp input, if one isn't provided
by the client.

This is to unify the request behavior, since some
graph mutations require timestamps (newer) and some
do not (older, backwards-compatible). The timestamp
of the mutation defaults to the subgraph server time
anyway if it is not provided, so there's not a benefit
to conditionally passing it through the proxy.