fix(v3proxy:send): fallback parsing for actions array
Previously we were assuming all query params were
still encoded strings; this is not true for some
GET request patterns. This PR first attempts to
parse actions array as JSON; if it fails, then
try decoding the string (assuming uri-encoded)
before reparsing as JSON.

In many cases it was fine to decode the already-
decoded string again, because the actions array
typically did not include special characters. But
since it's user input it's possible to say, make a tag
with a percent sign. That was causing the whole thing
to fail because it wasn't valid url encoding.

[POCKET-10048]