feat(exploreCorpus): process collection and corpus events
Set up terraform infrastructure for indexing new collections
and corpus items in elasticsearch cluster. Sets up rules for
forwarding relevant events to an SNS topic, which fans out to
SQS queue that is batch-consumed by a Lambda. This lambda
makes bulk index requests to elasticsearch cluster.

Contains the app code for the lambda indexer.

Updates the language-specific corpora indices -- remove a
duplicative ID field and add a parent_id field for collection
stories, which points to the parent collection.