fix(search:corpus): add status to corpus indices
The service now emits events for draft collections,
so we want to ensure we can differentiate between
published and draft collections when we return results.

Future enhancements might delete or discard this data,
but since it's possible to transition from Published
to Draft and there just isn't much data, it's not
expensive to just keep it around for now as we decide
on exactly how we want to model.