fix(search): update queries for os_2.13 upgrade (#618)
* fix(search): update queries for os_2.13 upgrade

Update localstack search cluster to be opensearch,
not elasticsearch. Note that the latest version we
could get locally is 2.11 but we are at 2.13 on AWS.

Remove 'type' (e.g. /_doc) from bulk query
endpoint.

Update score sort to explicitly include ordering,
rather than just passing the '_score' string.
Add a tiebreaker sort on 'item_id' for when _score
is identical among results.

Fix incorrect test assertions (perhaps due to some
changes in the scoring algorithm) -- manually verified
the code was functioning as expected given the results
from the search method.

[POCKET-10244]

* fix(corpus:search): update 'found' local in response

It's in the root field, not under _source anymore

* chore: update flaky test