fix(v3proxy): make item_id/url validator return bad user input error (#621)
* fix(v3proxy): accept numeric item_ids

Previously it was assumed only strings were
sent, but we can see that numeric IDs are sent
occasionally (which was causing an error when
calling string methods).

[POCKET-10281]

* fix(v3proxy): return 400 if missing identifier in actions

For /v3/send actions -- will fail all actions
in the batch if do not have identifiers.

Updating this to an InputValidationError changes
the response code to 400 from 500 and suppresses
the reporting in Sentry.

[POCKET-10281]

* chore(test): update test error string expectation