fix(corpus:index): fix collection urls and standardize casing
Makes requests to the parser for collections to generate
the records and identifiers used to federate them as Pocket Items.

Add some additional seeding to the test to reflect this.

Standardize casing for status (historically it's uppercase but
is now lowercase -- when we reprocess the dataset, standardize
on lowercase).

Create urls for collections rather than just putting the slugs
there; sometimes the slugs in historical data are full URLs, and
sometimes they are just slugs. Parse conditionally so when we
reprocess data they are all standardized.

[POCKET-10301]