feat(search:auth): enable logged-out corpus search
For native Pocket applications only.

Adds assertion that a user is logged in to
every other mutation and User field resolver.

This is an alternative to the router-based auth
directives we were trying to use, because of a
bug which causes the entire query to be rejected
if the unauthorized component references a variable.

https://github.com/apollographql/router/issues/5648

Since we probably will migrate to router-enforced auth
once this bug is fixed, this less elegant method requires
the least refactoring (vs. using the more complicated
auth models in other subgraphs)

[POCKET-10286]