fix(v3proxy:error): updates to error handling
Add more verbose logging for better deugging.

Consolidate async error handling patterns:
Allow all errors to bubble up to the route handler,
and wrap the route handler with a method which catches
rejected promises and sends them to error handling
middleware (otherwise resolve the function).

This might not be any different than the existing method
where the next calls are embedded in multiple try/catch
blocks of the async route handler, but I think this way
is less prone to developer error. I'm also wondering if
it's possible that the local test methods resolve promises
slightly differently than when the application is deployed?
Because we're seeing some unexpected unhandled errors
in Sentry, when no errors should be unhandled.

[POCKET-10354]