feat(corpus search): deduplicate syndicated articles
Remove the original article from the corpus
when a syndicated article is added. Assumes
that the original article is added to the
corpus first -- maybe there are some unlikely
edge cases which will result in duplicates, but
the tradeoff is worth it so that we don't have to
check every new write.

Choosing to delete rather than hide because the
syndicated article, if it expires or is removed,
will simply redirect to the original page.

[POCKET-10555]