feat(search:corpus): ingest remove approved item event
Delete a corpus item (not a collection) when a remove-approved-item
event is received.

Collections are special entities, punting handling that here.
We already hide them when the status is not 'published' so
it's not very risky. Need to connect with the content team
on how copies of collections are modeled in the corpus.

[POCKET-10325]