fix(list): control read/write connection in users meta
Users meta can be written to as a side effect
of certain v3 queries (e.g. tagslist). Previously
the writer was toggled on to be used for everything
to avoid a permissions error, but this resulted in
heavy load on the writer node. This update uses
the read-only user for read queries rather than
the writer or a dynamically set global. Since we
don't care about the return value of any changed
results on this table, we do not need to worry about
managing the connection to avoid issues with eventual
consistency (replication delay).