feat(export): asynchronous export service for user list (#819)
* feat(export): asynchronous export service for user list

An event-driven service for user list export, kicked off
by a mutation in list-api. To replace our synchronous
export process.

Exports a user's list in parts (10k rows each). These
are saved as csvs, which are zipped into an archive once
all export processes are completed.

When the export process is complete, the service emits
an event to the event bus, which is ingested by our
transactional email service. Users will be sent their
archive via a signed url, which expires after 14 days.
The archive itself expires after 30 days. If a user
requests another export in that 30 day period, they wilL
be sent a new link to the existing export. When it has
expired, a new export may be requested.

[POCKET-10669]

* chore: fix an unmerged head ref

* chore: fix some docstrings/remove comments

* chore: fix mismatches

* chore: lint

* chore: shorten sqs name